made pages folder into src diroctory and moved homepage and loading page in to the page folder
